refactor: extract @the-search-thing/api as shared HttpApi contract
Move schemas, domain errors, and the HttpApi definition into a browser-safe
workspace package so apps/web no longer imports backend server code.
Backend handlers and the web client both consume the same typed contract.
