+++ packages/api/src/domain/DocumentIndexError.ts
@@ -1,0 +1,8 @@
+import { Schema } from "effect";
+
+export class DocumentIndexError extends Schema.TaggedErrorClass<DocumentIndexError>()(
+  "DocumentIndexError",
+  {
+    message: Schema.String,
+  },
+) {}
