+++ packages/api/src/index.ts
@@ -1,0 +1,16 @@
+export { Api, HealthApi, IndexApi, SearchApi } from "./Api.js";
+export { DocumentIndexError } from "./domain/DocumentIndexError.js";
+export {
+  FileSearchError,
+  GrepModeSchema,
+  type GrepMode,
+} from "./domain/FileSearchSchemas.js";
+export {
+  ContentSearchItem,
+  ContentSearchResponse,
+  FileSearchItem,
+  FileSearchResponse,
+  HealthResponse,
+  IndexRunErrorItem,
+  IndexRunResponse,
+} from "./schemas.js";
