+++ packages/backend/src/search/FileSearchService.ts
@@ -1,15 +1,5 @@
-import { Context, Effect, Schema } from "effect";
-
-export const GrepModeSchema = Schema.Union([
-  Schema.Literal("plain"),
-  Schema.Literal("fuzzy"),
-  Schema.Literal("regex"),
-]);
-export type GrepMode = typeof GrepModeSchema.Type;
-
-export class FileSearchError extends Schema.TaggedErrorClass<FileSearchError>()("FileSearchError", {
-  message: Schema.String,
-}) {}
+import { Context, Effect } from "effect";
+import { FileSearchError, type GrepMode } from "@the-search-thing/api";
 
 export class FileSearchService extends Context.Service<
   FileSearchService,
